Updates to rainfall function locations and import statements
Moved several base rainfall functions to common, so that different
sub-modules can better share their imports.
